+++ libs/ui/svelte/components/Icon/InlineSVG.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { onMount } from "svelte";
   import type { ClassValue } from "svelte/elements";
+  import { fetchSVG, svgs } from "./InlineSVGState.svelte";
 
   interface Props {
     src: string;
@@ -10,34 +10,14 @@
 
   let { src, class: className, ...rest }: Props = $props();
 
-  let svgContent = $state<string | null>(null);
-  let error = $state<string | null>(null);
-
-  async function fetchSVG() {
-    try {
-      const response = await fetch(src);
-      if (!response.ok) {
-        throw new Error(`Failed to fetch SVG: ${response.statusText}`);
-      }
-      const svgText = await response.text();
-      svgContent = svgText;
-    } catch (err: any) {
-      error = err?.message;
-    }
-  }
-
-  onMount(() => {
-    fetchSVG();
-  });
+  let svgContent = $derived(svgs[src]);
 
   $effect(() => {
-    svgContent = null;
-    error = null;
-    fetchSVG();
+    fetchSVG(src);
   });
 </script>
 
-{#if !error && svgContent}
+{#if svgContent}
   <i {...rest} class={["inline-svg", className]}>
     {@html svgContent}
   </i>