+++ src/ui/svelte/flyouts/state/placement.svelte.ts
@@ -1,0 +1,73 @@
+import { invoke, SeelenCommand, SeelenEvent, subscribe, Widget } from "@seelen-ui/lib";
+import { lazyRune } from "libs/ui/svelte/utils";
+import { ConfigState } from "./config.svelte";
+import { PhysicalPosition } from "@tauri-apps/api/window";
+
+let showing = $state(false);
+
+const monitors = lazyRune(() => invoke(SeelenCommand.SystemGetMonitors));
+await subscribe(SeelenEvent.SystemMonitorsChanged, monitors.setByPayload);
+await monitors.init();
+
+const primaryMonitor = $derived.by(() => {
+  return monitors.value.find((m) => m.isPrimary) || monitors.value[0];
+});
+
+const widgetSize = lazyRune(() => Widget.self.webview.outerSize());
+await Widget.self.webview.onResized(widgetSize.setByPayload);
+await widgetSize.init();
+
+$effect.root(() => {
+  $effect(() => {
+    const monitor = primaryMonitor;
+
+    if (!monitor || !showing) {
+      return;
+    }
+
+    const placement = ConfigState.config.placement;
+    const padding = ConfigState.config.margin * monitor.scaleFactor;
+
+    const monitorWidth = monitor.rect.right - monitor.rect.left;
+    const monitorHeight = monitor.rect.bottom - monitor.rect.top;
+
+    const monitorCenterX = monitor.rect.left + monitorWidth / 2;
+    const monitorCenterY = monitor.rect.top + monitorHeight / 2;
+
+    const { width, height } = widgetSize.value;
+
+    let x: number, y: number;
+    if (placement === "left") {
+      x = monitor.rect.left + padding;
+      y = Math.round(monitorCenterY - height / 2);
+    } else if (placement === "top") {
+      x = Math.round(monitorCenterX - width / 2);
+      y = monitor.rect.top + padding;
+    } else if (placement === "right") {
+      x = monitor.rect.right - width - padding;
+      y = Math.round(monitorCenterY - height / 2);
+    } else {
+      x = Math.round(monitorCenterX - width / 2);
+      y = monitor.rect.bottom - height - padding;
+    }
+
+    Widget.self.webview.setPosition(new PhysicalPosition(x, y)).then(() => Widget.self.show(false));
+  });
+});
+
+export function setShowing(value: boolean) {
+  showing = value;
+
+  if (!value) {
+    Widget.self.hide();
+  }
+}
+
+export const Monitors = {
+  get all() {
+    return monitors.value;
+  },
+  get primary() {
+    return primaryMonitor;
+  },
+};